refac: cross asset racing league and pac man
