--- examples/arkanoid/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "arkanoid"
-version = "0.1.0"
-edition = "2021"
-publish = false
-description = "Arkanoid on-chain game using Cougr-Core ECS framework on Stellar Soroban"
-license = "MIT OR Apache-2.0"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-soroban-sdk = "25.1.0"
-cougr-core = { path = "../../" }
-
-[dev-dependencies]
-soroban-sdk = { version = "25.1.0", features = ["testutils"] }
-
-[profile.release]
-opt-level = "z"
-overflow-checks = true
-debug = 0
-strip = "symbols"
-debug-assertions = false
-panic = "abort"
-codegen-units = 1
-lto = true
-
-[profile.release-with-logs]
-inherits = "release"
-debug-assertions = true
