--- examples/asteroids/.gitignore
@@ -1,8 +1,0 @@
-# Rust's output directory
-target
-
-# Local settings
-.soroban
-.stellar
-
-/test_snapshots/
