--- examples/asteroids/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "asteroids"
-version = "0.0.0"
-edition = "2021"
-publish = false
-
-[lib]
-crate-type = ["lib", "cdylib"]
-doctest = false
-
-[dependencies]
-soroban-sdk = "25.1.0"
-cougr-core = { path = "../../" }
-
-[dev-dependencies]
-soroban-sdk = { version = "25.1.0", features = ["testutils"] }
-
-[profile.release]
-opt-level = "z"
-overflow-checks = true
-lto = true
-codegen-units = 1
-panic = "abort"
