--- examples/asteroids/src/Makefile
@@ -1,16 +1,0 @@
-default: build
-
-all: test
-
-test: build
-	cargo test
-
-build:
-	stellar contract build
-	@ls -l target/wasm32v1-none/release/*.wasm
-
-fmt:
-	cargo fmt --all
-
-clean:
-	cargo clean
