--- examples/battleship/.github/workflows/battleship.yml
@@ -1,99 +1,0 @@
-name: Battleship CI
-
-on:
-  push:
-    branches: [ main ]
-    paths:
-      - 'examples/battleship/**'
-  pull_request:
-    branches: [ main ]
-    paths:
-      - 'examples/battleship/**'
-
-env:
-  CARGO_TERM_COLOR: always
-
-jobs:
-  test:
-    name: Test Battleship Contract
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: examples/battleship
-
-    steps:
-    - uses: actions/checkout@v4
-
-    - name: Install Rust
-      uses: actions-rust-lang/setup-rust-toolchain@v1
-      with:
-        toolchain: stable
-        components: rustfmt, clippy
-
-    - name: Cache cargo registry
-      uses: actions/cache@v4
-      with:
-        path: ~/.cargo/registry
-        key: ${{ runner.os }}-cargo-registry-${{ hashFiles('**/Cargo.lock') }}
-
-    - name: Cache cargo index
-      uses: actions/cache@v4
-      with:
-        path: ~/.cargo/git
-        key: ${{ runner.os }}-cargo-index-${{ hashFiles('**/Cargo.lock') }}
-
-    - name: Cache cargo build
-      uses: actions/cache@v4
-      with:
-        path: examples/battleship/target
-        key: ${{ runner.os }}-cargo-build-target-${{ hashFiles('**/Cargo.lock') }}
-
-    - name: Check formatting
-      run: cargo fmt -- --check
-
-    - name: Run clippy
-      run: cargo clippy -- -D warnings
-
-    - name: Run tests
-      run: cargo test --verbose
-
-    - name: Build contract
-      run: cargo build --release --target wasm32-unknown-unknown
-
-  build:
-    name: Build WASM
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: examples/battleship
-
-    steps:
-    - uses: actions/checkout@v4
-
-    - name: Install Rust
-      uses: actions-rust-lang/setup-rust-toolchain@v1
-      with:
-        toolchain: stable
-        target: wasm32-unknown-unknown
-
-    - name: Install Stellar CLI
-      run: |
-        cargo install --locked stellar-cli --features opt
-
-    - name: Build optimized WASM
-      run: stellar contract build
-
-    - name: Check WASM size
-      run: |
-        WASM_SIZE=$(wc -c < target/wasm32-unknown-unknown/release/battleship.wasm)
-        echo "WASM size: $WASM_SIZE bytes"
-        if [ $WASM_SIZE -gt 1048576 ]; then
-          echo "Warning: WASM size exceeds 1MB"
-        fi
-
-    - name: Upload WASM artifact
-      uses: actions/upload-artifact@v4
-      with:
-        name: battleship-wasm
-        path: examples/battleship/target/wasm32-unknown-unknown/release/battleship.wasm
-        retention-days: 7
